requirements.txt
flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.0.0
PyMuPDF==1.23.7
pdf2image==1.16.3
pillow==10.0.1
google-generativeai==0.3.1
requests==2.31.0
tqdm==4.66.1
pathlib==1.0.1
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
seaborn>=0.12.0 